requirements.txt
requests==2.31.0
setuptools==68.2.2
scapy==2.5.0
keyring==24.2.0
python-crontab==3.0.0
environs==9.5.0
click==8.1.7
cython==3.0.2
pyinstaller==6.0.0
sqlitedict==2.1.0
python-nmap
paramiko
pyyaml
schedule
pysnmplib==5.0.21